Move scripts to scripts directory
