+++ web/scripts/rate_all.py
